+++ src/nti/site/localutility.py
@@ -12,6 +12,8 @@ __docformat__ = "restructuredtext en"
 logger = __import__('logging').getLogger(__name__)
 
 from zope import component
+from .utils import registerUtility
+from .utils import unregisterUtility
 
 def install_utility(utility, utility_name, provided, local_site_manager):
     """
@@ -35,7 +37,7 @@ def install_utility(utility, utility_name, provided, local_site_manager):
     # bases
 
     local_site_manager[utility_name] = utility
-    local_site_manager.registerUtility(utility, provided=provided)
+    registerUtility(local_site_manager, utility, provided=provided)
 
 def install_utility_on_registration(utility, utility_name, provided, event):
     """
@@ -67,9 +69,9 @@ def uninstall_utility_on_unregistration(utility_name, provided, event):
     child_component = local_site_manager[utility_name]
 
     looked_up = local_site_manager.getUtility(provided)
-    assert looked_up is child_component
+    assert looked_up is child_component, (looked_up, child_component)
 
-    local_site_manager.unregisterUtility(child_component, provided=provided)
+    unregisterUtility(local_site_manager, child_component, provided=provided)
     del local_site_manager[utility_name]
 
 def queryNextUtility(context, interface, default=None):
@@ -124,5 +126,6 @@ def queryNextUtility(context, interface, default=None):
         result = all_utilities[next_]
         if result is context:
             # in the GSM, we're querying for the GSM utility?
-            result = default
+            # Terminate with default to avoid an infinite loop
+            result = default # pragma: no cover
     return result